server/requirements.txt
# FastAPI and Server Essentials
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic[email]
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
bcrypt==3.2.2 # Pinned version for compatibility with passlib
python-multipart
redis

# --- AI & RAG Libraries ---

# LangChain Core and Integrations
# These replace the manual implementation libraries
langchain
langchain-community
langchain-openai
langchain-anthropic
langchain-huggingface

# Vector Store & Embeddings
faiss-cpu
sentence-transformers
tiktoken

# Document Loaders
pypdf

# PyTorch (CPU-only version for smaller size and faster builds)
# Use --extra-index-url so pip checks BOTH the default index and the PyTorch index
--extra-index-url https://download.pytorch.org/whl/cpu
torch
torchvision
torchaudio

scikit-learn